Integraate ModelMapper library and refactor DepartmentService Implementation class

+++ src/main/java/com/dineshmane/ems/serviceImpl/DepartmentServiceImpl.java
@@ -6,9 +6,9 @@ import com.dineshmane.ems.exception.DepartmentNotExistsException;
 import com.dineshmane.ems.repository.DepartmentRepository;
 import com.dineshmane.ems.service.DepartmentService;
 import lombok.AllArgsConstructor;
+import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -17,33 +17,29 @@ import java.util.stream.Collectors;
 public class DepartmentServiceImpl implements DepartmentService {
 
     private DepartmentRepository departmentRepository;
-
+    private ModelMapper modelMapper;
 
     @Override
     public DepartmentDTO addDepartment(DepartmentDTO departmentDTO) {
-        Department department = new Department();
-        department.setDepartmentName(departmentDTO.getDepartmentName());
-        department.setDepartmentDescription(departmentDTO.getDepartmentDescription());
-
+        Department department = modelMapper.map(departmentDTO, Department.class);
         Department savedDepartment = departmentRepository.save(department);
-
-        return new DepartmentDTO(savedDepartment.getId(), savedDepartment.getDepartmentName(), savedDepartment.getDepartmentDescription());
+        return modelMapper.map(savedDepartment, DepartmentDTO.class);
     }
 
     @Override
     public DepartmentDTO getDepartmentById(Long id) {
 
         Department department = departmentRepository.findById(id).orElseThrow(()-> new DepartmentNotExistsException("Department with given id is not exists"));
-        return new DepartmentDTO(department.getId(), department.getDepartmentName(), department.getDepartmentDescription());
+        return modelMapper.map(department, DepartmentDTO.class);
     }
 
     @Override
     public DepartmentDTO updateDepartment(DepartmentDTO departmentDTO) {
-        Department department = departmentRepository.getDepartmentById(departmentDTO.getId()).orElseThrow(()-> new DepartmentNotExistsException("Department with given doesn't exists"));
+        Department department = departmentRepository.findById(departmentDTO.getId()).orElseThrow(()-> new DepartmentNotExistsException("Department with given doesn't exists"));
         department.setDepartmentName(departmentDTO.getDepartmentName());
         department.setDepartmentDescription(departmentDTO.getDepartmentDescription());
         Department savedDepartment = departmentRepository.save(department);
-        return new DepartmentDTO(savedDepartment.getId(), savedDepartment.getDepartmentName(), savedDepartment.getDepartmentDescription());
+        return modelMapper.map(savedDepartment, DepartmentDTO.class);
     }
 
     @Override
@@ -51,14 +47,14 @@ public class DepartmentServiceImpl implements DepartmentService {
         List<Department> departments = departmentRepository.findAll();
 
         return departments.stream()
-                .map(d->(new DepartmentDTO(d.getId(), d.getDepartmentName(), d.getDepartmentDescription())))
+                .map(d->(modelMapper.map(d, DepartmentDTO.class)))
                 .collect(Collectors.toList());
     }
 
     @Override
     public void deleteDepartmentByid(Long id) {
+        if(!departmentRepository.existsById(id)) throw new DepartmentNotExistsException("Department with given id is not present");
         departmentRepository.deleteById(id);
     }
-
 
 }
